# api.yaml
swagger: "2.0.0"
info:
  title: Test API Documentation
  description: Api Documentation for Test App Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:     
...
  /v1/admin/filterStudentResult/:page/:limit:
    get:
      tags:
        - name: Students
      summary: Get All Board By filter
      description: APi to get all the Board by filter
      parameters:
        - name: searchString
          in: query
          type: string
        - name: page
          in: path
          type: string
        - name: limit
          in: path
          type: string
        - name: status
          in: query
          type: string
        - name: startDate
          in: query
          type: string
        - name: endDate
          in: query
          type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
...
  /v1/admin/getAllStudentsReferral/{offset}/{limit}:
    get:
      tags:
        - name: Referral
      summary: Student With Referral 
      description: Get all the students with the referral 
      parameters:
        - in: path
          name: offset
          type: integer
          required: true
        - in: path 
          name: limit 
          type: integer
          required: true
      responses:
        200:
          description: OK
  /v1/admin/deleteStudent/{studentID}:
    delete:
      tags:
        - name: Student
      summary: Delete student 
      description: APi that deleted student 
      parameters:
        - name: studentID 
          in: path
          type: string 
          required: true 
      responses:
        200:
          description: OK
  /v1/admin/getAllStudentsWithoutReferral/{offset}/{limit}:
    get:
      tags:
        - name: Student
      summary: Student With without Referral 
      description: lorem is the one
      parameters:
        - name: offset
          in: path
          required: true 
          type: integer
        - name: limit
          in: path
          required: true 
          type: integer
      responses:
        200:
          description: OK
  /v1/admin/getAllStudents/{studentID}:
    get:
      tags:
        - name: Student
      summary: STudent By ID 
      description: Will get all the bookings of the specified Student imageID
      parameters:
        - name: studentID
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
  /v1/admin/getAllStudents/{offset}/{limit}:
    get:
      tags: 
        - name: Student
      summary: Getting ALl Students
      description: Api will get all Students      
      parameters:
        - in: path
          name: offset
          required: true
          type: integer
        - in: path
          name: limit
          type: integer
          required: true
      responses:
        200:
          description: OK
  # /v1/customer/bookings/{barberID}/{status}:
  #   get:
  #     tags:
  #       - name: Bookings for the Barber
  #     summary: Get Bookings Of Barber by status
  #     description: Will get all the bookings of the specified Barber id and status
  #     parameters:
  #       - name: barberID
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #       - name: status
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       200:
  #         description: OK
  # /v1/customer/bookingsGreater/{barberID}/{status}:
  #   get:
  #     tags:
  #       - name: Bookings for the Barber
  #     summary: Get Bookings Of Barber greater Than date.now()
  #     description: Will get all the bookings of the specified Barber id and date
  #     parameters:
  #       - name: barberID
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #       - name: status
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       200:
  #         description: OK
  # /v1/customer/bookingsLesser/{barberID}/{status}:
  #   get:
  #     tags:
  #       - name: Bookings for the Barber
  #     summary: Get Bookings Of Barber less than Date.now()
  #     description: Will get all the bookings of the specified Barber id and date
  #     parameters:
  #       - name: barberID
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #       - name: status
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       200:
  #         description: OK
  # /v1/customer/bookingsDate/{barberID}/{date}:
  #   get:
  #     tags:
  #       - name: Bookings for the Barber
  #     summary: Get Bookings Of Barber of Particular Date
  #     description: Will get all the bookings of the specified Barberif and date
  #     parameters:
  #       - name: barberID
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #       - name: date
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       200:
  #         description: OK
...
  /v1/admin/createStudent:
    post:
      tags:
        - name: Student
      summary: Adding Students
      description: Api will be used to add students in the db
      consumes:
        application/json
      produces:
        application/json
      parameters:
        - in: body
          name: student
          schema:
            type: object
            properties:
              userName: 
                type: string 
              email: 
                type: string 
              password: 
                type: string
      responses:
        200:
          description: OK
...
  /v1/admin/deleteTeacher/{teacherID}:
    delete:
      tags: 
        - name: Teacher
      summary: Delete Teacher
      description: This api will delete the teacher 
      parameters:
        - name: teacherID
          in: path
          type: string 
          required: true
      responses:
        200:
          description: OK
...
  /v1/admin/createTeacher:
    post:
      tags: 
        - name: Teacher
      summary: Adding Teachers
      description: Adding teachers to the Database of auth By the Admin
      consumes:
        application/json
      produces:
        application/json
      parameters:
      - in: body
        name: User
        schema:
          type: object
          properties:
            userName:
              type: string 
            email:
              type: string
            password:
              type: string
            commisionPercent:
              type: string
      responses:
        200:
          description: OK